extract the draw plugin installation
